requirements.txt
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
blis==0.7.6
Brlapi==0.7.0
catalogue==2.0.6
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.4
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cymem==2.0.6
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
en-core-web-sm==3.2.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
filelock==3.4.2
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
keyring==18.0.1
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
murmurhash==1.0.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.10
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathy==0.6.1
pdfminer.six==20211012
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.4.1
preshed==3.0.6
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.6.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.8.2
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
systemd-python==234
terminado==0.13.3
testpath==0.6.0
thinc==8.0.15
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typer==0.4.0
typing-extensions==4.1.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.0
wadllib==1.3.3
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
xkit==0.0.0
zipp==3.7.0
